Add [make error-predecessors].
This entry in Makefile.menhir computes and prints the set of all (non-%inline)
non-terminal symbols whose FOLLOW set contains the token [error]. This can be
used to determine which symbols should carry an [%on_error_reduce] declaration
when moving to the new scheme for reporting syntax errors.

This entry can be removed (in the future) once the [error] token has been
removed from the grammar.