Add %on_error_reduce declarations to ensure full reduction before error.
A manual inspection of the error states suggests that, by adding these
17 declarations, we ensure that every possible reduction is performed
before attempting to explain an error. In other words, the remaining
error states are states where no reduction can be performed.

For reasons that I have not investigated in detail, this changes the
syntax error message that is obtained in three sentences in the test
suite:

  implementation: LETOP UNDERSCORE BAR UNDERSCORE WITH
  implementation: LETOP UNDERSCORE COLONCOLON UNDERSCORE WITH
  implementation: LETOP UNDERSCORE COMMA UNDERSCORE WITH

The previous message included "pattern expected", where the new message
does not. One can argue that the old message was intuitively wrong:
after `let* _, _` one does not expect a pattern.